channid as uint16_t instead of int

+++ tests/includes/test_only/front/fake_front.cpp
@@ -38,10 +38,9 @@ const CHANNELS::ChannelDefArray & FakeFront::get_channel_list() const
     return d->cl;
 }
 
-void FakeFront::add_channel(CHANNELS::ChannelNameId name_id, uint32_t flags, int chanid)
+void FakeFront::add_channel(CHANNELS::ChannelNameId name_id, uint32_t flags, uint16_t chanid)
 {
     this->d->cl.push_back({name_id, flags, chanid});
-
 }
 
 FakeFront::ResizeResult FakeFront::server_resize(ScreenInfo screen_server)
